move files to src and lib
